Database/testing
route functionality and very basic validation is done. Next stop: auth and security

+++ glaring-api/db/migrate/20200607231307_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[6.0]
       t.string :name
       t.string :password
       t.string :email
-      t.numeric :phone_number
+      t.bigint :phone_number
       t.string :avatar
       t.timestamps
     end
